move containers into their own subfolder
