+++ src/org/ecs160/a2/ui/SummaryAll.java
@@ -5,6 +5,10 @@ import java.util.List;
 import com.codename1.ui.layouts.BoxLayout;
 
 import org.ecs160.a2.models.Task;
+import org.ecs160.a2.ui.containers.UpdateableContainer;
+import org.ecs160.a2.ui.containers.TaskContainer;
+import org.ecs160.a2.ui.containers.SizeContainer;
+import org.ecs160.a2.ui.containers.StatsContainer;
 import org.ecs160.a2.utils.AppConstants;
 import org.ecs160.a2.utils.UIUtils;
