+++ src/org/ecs160/a2/ui/SummarySize.java
@@ -8,6 +8,9 @@ import com.codename1.ui.layouts.BoxLayout;
 import com.codename1.ui.spinner.Picker;
 
 import org.ecs160.a2.models.Task;
+import org.ecs160.a2.ui.containers.UpdateableContainer;
+import org.ecs160.a2.ui.containers.TaskContainer;
+import org.ecs160.a2.ui.containers.StatsContainer;
 import org.ecs160.a2.utils.AppConstants;
 import org.ecs160.a2.utils.UIUtils;
